Vast Improvement
Improved IIMC's output by using "numSimulations % b.getDeck().size()" to determine which card to pass, rather than random.
I believe this improvement is caused by a more normal distribution of cards being used.

On a side-note, Main was updated to allow you to pass "-s=NUM" to change the start seed to NUM, and/or "-g=NUM2" to set the number of games to NUM2